add test case for RoundTripBalancer
